+++ internal/api/token_refresh.go
@@ -211,12 +211,12 @@ func (a *API) RefreshTokenGrant(ctx context.Context, w http.ResponseWriter, r *h
 				}
 			}
 
-			if terr = models.NewAuditLogEntry(r, tx, user, models.TokenRefreshedAction, "", nil); terr != nil {
+			if terr = models.NewAuditLogEntry(config.AuditLog, r, tx, user, models.TokenRefreshedAction, "", nil); terr != nil {
 				return terr
 			}
 
 			if issuedToken == nil {
-				newToken, terr := models.GrantRefreshTokenSwap(r, tx, user, token)
+				newToken, terr := models.GrantRefreshTokenSwap(config.AuditLog, r, tx, user, token)
 				if terr != nil {
 					return terr
 				}
@@ -276,7 +276,7 @@ func (a *API) RefreshTokenGrant(ctx context.Context, w http.ResponseWriter, r *h
 				return err
 			}
 		}
-		metering.RecordLogin("token", user.ID)
+		metering.RecordLogin(metering.LoginTypeToken, user.ID, nil)
 		return sendJSON(w, http.StatusOK, newTokenResponse)
 	}